requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.5
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-jet==1.0.8
django-js-asset==1.2.2
django-robots==4.0
django-simple-math-captcha==1.0.9
django-sitemaps==1.1.1
django-user-tracking==0.4.11
lxml==4.5.0
Pillow==7.1.1
pkg-resources==0.0.0
pytz==2019.3
sentry-sdk==0.14.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
